builder: convert from full patch to module
The builder used to have to be a patch because the peripheral_processors
dicitonary was not accessible, but that was changed in
ARTIQ's 52112d54f9c052159b88b78dc6bd712abd4f062c.
This refactors the patch into a proper, fairly clear module.

This module extends artiq/gateware/targets/kasli_generic.py,
so the syntax is the same with added Entangler functionality.
This isn't compatible with other gateware extensions, but that could
be changed in the future.